Move the fs_queue implementation to the top level
